heap: the driver that answers a guest ABI question is a host capability, not a rail
`runtime/heap_query.rs` decodes `CmdHeapTextureSizeAndAlign` off the wire and
then, in the same file, opened an `MTLDevice` and called
`heapTextureSizeAndAlign:` — a neutral runtime module linking the Metal
framework directly, past `backend/` entirely.

The driver call moves to `backend/heap_placement.rs`. What it is *not* is the
point of the module doc:

- Not a `Backend` method. The guest is asking what Apple's driver requires to
  place a texture in a heap, which is a term of the guest ABI. The arm64 macOS
  pathway runs one guest stream through Metal and through MoltenVK without a
  rebuild, and a heap layout that differed between them would be this device
  answering a guest contract with a host implementation detail.
- Not part of `backend::metal`. That module is gated on
  `feature = "backend-metal"`; this call has always been gated on
  `target_os = "macos"` alone, because the question is whether a driver exists
  to be asked. Filing it under the Metal rail would silently make the shipping
  macOS Vulkan pathway start refusing a query it answers today.

`runtime/heap_query.rs` keeps the wire — request decode, descriptor, reply
encoding, refusal vocabulary — and the one live test that carries a real
fixture's descriptor through to the driver stays with the decode it is about.

The neutral runtime and model now name a rail in exactly two places outside
`cfg(test)`: the `MetalBackend(metal::error::Status)` variants of `EncodeStatus`
and `ComputeStatus`.

Verified: feature matrix all arms (counts unchanged), 2767 passed / 0 failed,
clippy unchanged on all three arms.
